Update Pickup
The EasyPost pickup is created with pickup_rates using which it is purchased. When the pickup is initially added, it is in unknown state.
When pickup is purchased with EasyPost buy pickup API using carrier and service, the state is moved to scheduled.

The proposed modification aims to purchase the newly added pickup with the cheapest pickup rate, adds pickup information to metadata.
The new test case verifies the working when the shipping_method is not scheduled for pickups.